Block possibility of going back to login from bottom_navigation screens(owner and user). Typo in owner_bottom_navigation
