+++ lib/screens/authentication_flow/login_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:hotel_manager/screens/owner_flow/owner_botton_navigation.dart';
+import 'package:hotel_manager/screens/owner_flow/owner_bottom_navigation.dart';
 import 'package:hotel_manager/screens/authentication_flow/sign_up_choice_screen.dart';
 import 'package:hotel_manager/screens/user_flow/user_bottom_navigation.dart';
 import 'package:hotel_manager/services/user_service.dart';
@@ -52,13 +53,27 @@ class _LoginViewState extends State<LoginView> {
     }
   }
 
+  void _closeKeyboard() {
+    FocusScopeNode currentFocus = FocusScope.of(context);
+    if (!currentFocus.hasPrimaryFocus) {
+      currentFocus.unfocus();
+    }
+    SystemChannels.textInput.invokeMethod('TextInput.hide');
+  }
+
   void _navigateToLoggedUser() {
+    _closeKeyboard();
+    emailController.text = "";
+    passwordController.text = "";
     Navigator.of(context).push(MaterialPageRoute(
       builder: (context) => const UserBottomNavigationView(),
     ));
   }
 
   void _navigateToLoggedOwner() {
+    _closeKeyboard();
+    emailController.text = "";
+    passwordController.text = "";
     Navigator.of(context).push(MaterialPageRoute(
       builder: (context) => const OwnerBottomNavigationView(),
     ));
