+++ lib/screens/start_screen.dart
@@ -4,7 +4,7 @@ import 'package:hotel_manager/screens/authentication_flow/login_menu_screen.dart
 import 'package:hotel_manager/screens/user_flow/user_bottom_navigation.dart';
 
 import '/services/user_auth.dart';
-import 'owner_flow/owner_botton_navigation.dart';
+import 'owner_flow/owner_bottom_navigation.dart';
 
 class StartView extends StatefulWidget {
   const StartView({Key? key});
